requirements.txt
# Core ML and Deep Learning Libraries
jax-ai-stack>=0.1.0

# Hyperparameter Optimization
optuna>=3.0.0

# Data Processing and Analysis
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Financial Data
yfinance>=0.2.0
backtesting>=0.3.0
alpha-vantage>=2.3.1
polygon-api-client>=1.12.0

# Visualization and Plotting
matplotlib>=3.5.0
seaborn>=0.11.0

# Rich Terminal Output
rich>=12.0.0

# Progress Bars
tqdm>=4.62.0

# TensorBoard Integration
tensorflow>=2.8.0
tensorboard>=2.8.0

# Image Processing (for TensorBoard)
Pillow>=9.0.0

# System Monitoring
psutil>=5.8.0

# Checkpoint Management
orbax-checkpoint>=0.4.0
etils>=0.6.0

# Development and Testing (optional)
# pytest>=6.0.0
# pytest-cov>=2.0.0
# black>=21.0.0
# flake8>=3.8.0
# mypy>=0.800 

sentence-transformers>=2.2.2 
tf-keras>=2.18.0
